Final Commit
Ready for Review. Renamed the file names to their proper names. Added the scripts and a css file to the html. Added an event listener to the ws_cloud.js script. That listener gets the text content of the speech and sorts it by the most used word. The top three most used word are in pink and are large. The words are randomly rotated along with their size being changed based upon their position in how many words there are. Added in comments to the code.